don't use tokens for build new ast for inline (#532)
I studied in detail and came to the conclusion that this is the only working solution at the moment, not breaking the code.

Why?
Inline nodes can be in branch of control structures and it is very hard implement. (example inline node can be inside `else if` or `else`). Introduce new ast nodes (or andd based on tokens) is very bad solution (i vote doesn't do this never), many nodes can contain branches like `test`, `esle`, `testExpr`, `arguments` and etc, we will have to run through them - it is a lot of lines of code and a lot of bugs. Better wait fix in parses (or fix it and send PR).